refactor: repackage ocm client
